requirements.txt
# Core Flask
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
Flask-Migrate==4.0.5

# Database
psycopg2-binary==2.9.9
supabase==2.12.0

# AI/ML
google-genai==1.3.0
anthropic==0.7.10

# Utilities
python-dotenv==1.0.0
requests==2.31.0
pydantic==2.5.2
PyJWT==2.8.0
python-dateutil==2.8.2
ujson==5.8.0

# Production Server
gunicorn==21.2.0
greenlet==3.0.3

# Optional (if needed)
# ollama==0.1.6
# transformers==4.35.0
# torch==2.1.0